Disable syncing of Feature and Biome registries
Resolves https://github.com/MinecraftForge/MinecraftForge/issues/7451

Basically, Forge clients without a biome/feature mod cannot connect to a server with the mod. Vanilla clients can connect to the server with the mod.

The reason for this is Minecraft will send the necessary worldgen data from the server to the client which is how worldgen datapack with custom biomes and such will work perfectly fine entirely server side. Forge, however, forces the biome and feature registry to be exactly the same when a Forge client tries to connect to a server. This is not need and this PR turns off the syncing for those two registries to better match vanilla behavior.

I tested 36.0.14 client connecting to this PR's server and tested this PR's client connecting to 36.0.14 server. Both without worldgen stuff and it works both ways across the version difference safely.

I tested a 36.0.14 client with worldgen mod connecting to this PR's server and it allows the connection properly which makes sense as worldgen mods are useless on client.

Then lastly, I connected the 36.0.14 client to this PR's server with the test mod on and register worldgen stuff on the server. Client was able to connect just fine.

Of course, the 36.0.14 server with worldgen mods rejected this PR's client as the server is the Forge without this PR.

No major issues came up that I can tell from mismatch Forge versions with the testing I done above