Only fire LootingLevelEvent if there is a killer entity
This fix makes it so that the LootingLevelEvent only fires if there is a
killer entity, since looting without a killer makes little sense. Also
adds the Nullable annotation to LootingLevelEvent.getDamageSource() to
more clearly communicate to modders that the DamageSource can be null in
LootingLevelEvent.

See issue #7714 for a more in depth discussion about the origin of this
pull request.