added test mod to verify the PR works
Adds the Bastion spacing to Overworld and End. And will also remove Bastion spacing from Nether and the Custom dimension that is a clone of the Overworld.

Expectation: Bastion spawns in only Overworld and End.

Result if PR #7777 fails: Bastion only spawns in End and no other dimension.

Test result: expectation was met and the PR does indeed work.