It's intended!
Replace
`This could be an intended override, but in most cases indicates a broken mod.`
with
`This could be a broken mod, but in most cases indicates an intended override.`
I've seen this message over ten thousand times, and it has never actually been a sign of a broken mod. This has also confused some of my friends who were new to modding.
You may also want to consider suppressing this warning, as on most modpacks(tested on Omnifactory) this one message takes up a whopping **37%** of the log. I couldn't find where that would be handled.