Remove errorneous patch in ClientPlayNetHandler
This patch prevents the buffer in the custom payload packet handling
method (which is a buffer containing a copy of the data from the packet)
from being released. Removing this allows the buffer to be released as normal.

From the context of when this patch was added, it seems this was added
to another try-catch for "MC|TrList" packets, but was then misapplied
elsewhere, presumably because it was unclear the patch applied for
"MC|TrList" packets.