Make WorldEvent.Load run for Overworld before chunk creation like other dimensions
I was doing some testing by removing structures from a dimension by removing them from the world's chunk generate in WorldEvent.Load. What I noticed is while doing that worked perfectly for any dimension, Overworld was an exception where the structure still spawned around the world origin but my change seemed to apply past origin area (200 block radius around origin).

Looking at the patch for WorldEvent.Load, I noticed the event is added to a for loop that deliberately skips over the overworld inside but the hook is after the check. Meaning the Overworld's stuff was defined further above and when looking higher, I saw setInitialSpawn method being called after the Overworld is setup.

I believe setInitialSpawn is creating chunks around the spawn of the Overworld which causes WorldEvent.Load to fire when the overworld has already genned some chunks as is in a different state than the rest of the dimensions.

So this pr is to move the old WorldEvent.Load event into the loop's if statement to exclude the overworld there. Then make a second hook for the WorldEvent.Load event that is placed before setInitialSpawn method. That way when the event fires, modders know the overworld is not at a different stage or loading than other dimensions. Just something for better consistency.

Also, someone forgot to put a license header on their file so that's slipped into this pr to help. but I can remove this extra license file thing if preferred