Remove UNKNOWN removal reason with REMOVED
As the UNKNOWN reason serves little purpose other than currently a
placeholder value when calling the original setRemoved, it was decided
that it should be removed and be replaced with the REMOVED reason.

This means that callers of the original setRemoved() now signal that
the block entity is removed permanently. However, it is noted that this
would only be for existing mods; as the original setRemoved() is
deprecated, mods shouldn't be using it and should prefer the one with
the explicit removal reason.