Implement smarter detection of test sources
Instead of relying on the presence of the word `test` in the paths for
the test source set, the paths are now taken from the MOD_CLASSES
environment property, filtered by the configured 'mod ID' from the
system property `forge.test.id`. This property is set to the name of
the mods sources configured in the run configuration (currently,
`tests`).

This also modifies MinecraftLocator to exclude the test sources (as
configured above) from its scanning, avoiding a modloading warning on
startup due to the lack of `mods.toml` in the test source set resources
directory.