Add check to load only in Forgedev environments
CommonDevLaunchHandler is the common superclass for all dev launch
targets, userdev and Forgedev, while CommonUserdevLaunchHandler is the
common superclass for all userdev launch targets.

Doing an instanceof here based on those two facts is less fragile than
inspecting the names of the launch target using the naming patterns of
the launch handlers.