Improve invalid comment check to use proper path
The previous position of the comment check being in the `comment` method meant that the logged config path is incorrect.

This is because the actual config path is constructed from the current
path and the path passed in to one of the `define*` methods, while the
comment check at its previous position only has access to the current
path (which is the current category containing the value).

The comment check is now changed to flip an invalid comment flag. This
is then checked later when defining a value or creating a config
category, which has the full actual path of the value/category.