Call isEdible on stack in hand than on current item instance
Before the addition of the ItemStack sensitive #getFoodProperties
method, both #isEdible and (insensitive) #getFoodProperties were called
on the current item instance. However, after the addition, it
is now possible that #isEdible is called on a different item instance
than (sensitive) #getFoodProperties.

This fixes that potential issue by calling #isEdible on the stack in
hand rather than the current item instance, thus matching the stack used
 for (sensitive) #getFoodProperties.

Fixes #8602