Add User-Agent header to requests made by the update checker
Currently the update checker sends requests with the default Java HTTP client `User-Agent` header - `Java-http-client/<Java version>`. This is rather unhelpful for a couple reasons, principally that it's so generic that it kind of goes against the purpose of a user agent.

This PR changes the user agent header to make it clear that it's a request being made by Forge.

Example: `MinecraftForge/1.19-41.0.99 Java/17.0.1`