Allow mods to specify shader import namespace
Previously any included shaders would be hardcoded to search using the
`minecraft` namespace, limiting options for organisation options for
modders. This change allows modders to specify which namespace that
their shader imports should be loaded from, by prefixing their import
with the desired mod namespace.

To be completely honest, I am not a huge fan of this approach, but it
appears to be the least invasive, and also avoids a double-lookup
where we attempt to search the mod's namespace before falling back to
the minecraft namespace.