Compute additive/subtractive sets for intuitivity
This changes a good portion of the internals of TagLoader to compute the
entire closure of additive/subtractive tag entries, in order to return
the "most intuitive" thing.

As implemented, any entry in the 'remove' set takes precedence, even if
it is added back by another mod/datapack/etc. As an example:
```
// a/data/minecraft/tags/blocks/logs.json
{ "values": [], "replace": false, "remove": ["minecraft:oak_log"] }
// b/data/minecraft/tags/blocks/logs.json
{ "values": ["minecraft:oak_log"] }
```
The result here, regardless of load order, is that the `minecraft:logs`
tag does *not* include `minecraft:oak_log`.