Add support for custom chunk render types
This is the minimal set of changes needed to allow registration of custom
render types to be used when rendering static level geometry. I've
loosely based these changes on #8331 (such as including its mipmap fixes)
but does not include many of the other changes made in that PR, such as:
- Using the TierSortingRegistry to sort and manage RenderTypes
- Supporting transparent/tripwire geometry - this is because I expect
  that most developers would want the translucent sort to be applied for
  geometry that is rendered around the same time as translucent geometry
  and that would require a larger set of changes to decide whether to
  enable the translucent sort or not without harming performance too
  much
- A configuration flag to use the vanilla codepath: the Forge codepath is
  always enabled as disabling it will cause holes in the world
- The additional types for registering a callback before rendering chunk
  geometry used for uploading uniforms and the like
- Updating MultiLayerModel as it has been replaced by CompositeModel,
  which I believe should already handle these changes if modders are
  registering their RenderType-s properly.

As with the other PR, the additional runtime cost is roughly the same,
scaling with the same parameters.

Included is a test mod which registers a block which uses a shader to
render each face as its normals. It can be accessed using the setblock
command.