[mdk] Add a starter Github action workflow file
Just a small template that helps modders that choose to open-source their mods vet their
and others' changes.  All it does is validate that no one tampered with the gradle wrapper
(this is a possible attack vector that has been exploited before, iirc), and that the mod
compiles.

I think it'd be good to make gametests run by default too, but build.gradle says that
having no tests crashes the run config, so I didn't add one yet.