Support dependency requirement type
I am making an abstraction layer and in MC Forge discord itself some interest has been expressed for supporting the NeoForge style dependency format. Here is an implementation which will check if it is required and if so set to mandatory, optional is optional so is always false, though this does not have a warning for discouraged but still runs the mod anyhow. This is mostly a play at trying to make Neo mods.toml files parse correctly with the existing system rather than replacing it with Neo's.