Retain game event identifier in Forge event
As currently implemented in 1.21, Forge strips out the identifying information from the game event holder when creating a VanillaGameEvent object. This makes it impossible to tell which game event was fired. This change causes the VanillaGameEvent object to transmit the fully identified holder so that it can be properly handled by listeners.