Add `#forge:chorus_additionally_grows_on`
Allows modders to define blocks, aside from just `minecraft:end_stone`, for chorus plants to grow on. This is often used by modders who create new biomes for The End and have their own variations of the standard End Stone block.

The main reason why this tag is to add additional blocks rather than to replace the original check for End Stone is because Forge 1.20.6 has already been stabilized. Mods that add to the End are known to use hacks on the method call that checks for End Stone in the relevant classes. So, these are kept but the check for the new tag is tagged along with it.

Vanilla has a tag similar to this, `#minecraft:azalea_grows_on`, for the Azalea flowers feature.