Seal events that only have `@ApiStatus.Internal` constructors
Adding `final` to these event classes allows for additional optimisations to be made by both the JVM and EventBus (such as EventBus#85). Partial backport of #10515.

It's not possible to extend these events without calling the constructor that is marked as internal, therefore extending was considered for internal use only and should be made `final`.